add carrier policy / solving tests

+++ app/controllers/api/v2/admin/carriers_controller.rb
@@ -1,18 +1,19 @@
 class Api::V2::Admin::CarriersController < Api::V2::Admin::ContactsController
 
   def index
-    carriers = Carrier.all
+    carriers = policy_scope [:admin, Carrier]
     paginate json: carriers.order(:id), status: 200
   end
 
   def show
     carrier = Carrier.find(params[:id])
+    authorize [:admin, carrier]
     render json: carrier, status: 200
   end
 
   def create
     carrier = Carrier.new(category_params)
-
+    authorize [:admin, carrier]
     if carrier.save
       update_contact(carrier)
       render json: carrier, status: 201
@@ -23,7 +24,7 @@ class Api::V2::Admin::CarriersController < Api::V2::Admin::ContactsController
 
   def update
     carrier = Carrier.find(params[:id])
-
+    authorize [:admin, carrier]
     if carrier.update(category_params)
       update_contact(carrier)
       render json: carrier, status: 200
@@ -34,6 +35,7 @@ class Api::V2::Admin::CarriersController < Api::V2::Admin::ContactsController
 
   def destroy
     carrier = Carrier.find(params[:id])
+    authorize [:admin, carrier]
     carrier.destroy
     head 204
   end
@@ -41,6 +43,6 @@ class Api::V2::Admin::CarriersController < Api::V2::Admin::ContactsController
   private
 
   def category_params
-    params.permit(:name, :federal_registration, :state_registration, :kind, :description, :active)
+    params.permit(policy([:admin, Carrier]).permitted_attributes)
   end
 end
